+++ src/features/XIT/ACT/action-steps/TRANSFER_MATERIALS.ts
@@ -1,5 +1,5 @@
 import { act } from '@src/features/XIT/ACT/act-registry';
-import { serializeStorage } from '@src/features/XIT/ACT/actions/mtra/utils';
+import { serializeStorage } from '@src/features/XIT/ACT/actions/utils';
 import { fixed0 } from '@src/utils/format';
 import { changeInputValue, clickElement, focusElement } from '@src/util';
 import { materialsStore } from '@src/infrastructure/prun-api/data/materials';